5 - Streamlit/requirements.txt
altair==5.0.1
GitPython==3.1.43
imbalanced-learn==0.12.4
numpy==1.23.5
pandas==2.0.3
matplotlib==3.7.0
requests==2.28.1
scikit-learn==1.3.2
seaborn==0.12.2
shap==0.42.0
lazypredict==0.2.12
lightgbm==4.5.0
Pillow==9.4.0
plotly==5.11.0
streamlit==1.25.0
streamlit-shap==1.0.2
xplotter==0.0.6
xgboost==2.1.1